requirements.txt
# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
playwright>=1.40.0
lxml>=4.9.0

# Async Processing
aiohttp>=3.9.0
asyncio-mqtt>=0.16.0

# Data Processing
pandas>=2.1.0
pydantic>=2.5.0

# Azure Integration
azure-servicebus>=7.11.0
azure-storage-blob>=12.18.0
azure-identity>=1.15.0

# Proxy & Anti-Detection
fake-useragent>=1.4.0
cloudscraper>=1.2.71

# Logging & Monitoring
python-json-logger>=2.0.0
structlog>=23.2.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Utilities
schedule>=1.2.0
tenacity>=8.2.0
